Moved the test to existing files rather than adding new ones
